extract queryFromAsterixDB
